+++ client/src/pages/Sub/Cart/DeliveryFee.tsx
@@ -1,0 +1,42 @@
+import {
+  DeliveryFeeBar,
+  DeliveryFeeBarFull,
+  DeliveryFeeBox,
+  DeliveryFeeText,
+  DeliveryFeeTextBox,
+} from "./styles/DeliveryFee.style";
+import { SecondMediumText } from "./styles/Cart.style";
+
+function DeliveryFee() {
+  return (
+    <DeliveryFeeBox>
+      <DeliveryFeeTextBox>
+        {sumPrice ? (
+          <>
+            <DeliveryFeeText>
+              {sumPrice >= 30000 ? "무료배송" : 30000 - sumPrice}
+            </DeliveryFeeText>
+            <SecondMediumText>
+              {sumPrice >= 30000 ? null : "원 추가시 무료배송"}
+            </SecondMediumText>
+          </>
+        ) : (
+          <>
+            <SecondMediumText>30000원 추가시 무료배송</SecondMediumText>
+          </>
+        )}
+      </DeliveryFeeTextBox>
+      <DeliveryFeeBar>
+        {sumPrice ? (
+          <>
+            <DeliveryFeeBarFull priceSize={Number(sumPrice)}>
+              1
+            </DeliveryFeeBarFull>
+          </>
+        ) : null}
+      </DeliveryFeeBar>
+    </DeliveryFeeBox>
+  );
+}
+
+export default DeliveryFee;
